Terrain: Introduce Chunk Pool
Chunk Pool is a helper-wrapper for a map of Chunks. Its main purpose
is to memory-manage the Chunks and reuse these which were already
generated by us.

So far, Chunk Pool keeps in memory all saved chunks. In the future
this must be changed - ex. chunks which will exceed 2x visible radius
shall be discarded from memory.

Verification:
    Build, run, fly around. Notice that when going back the same way
    to 0, 0 Chunk, there should be no Chunks generated.

Change-Id: I63784c2219a34f80a5a15d8dd3957680299697d0