Terrain: Picking with ray-OBB intersection
Picking implementation with ray OBB intersection. Every click with RMB
enables flag, that in next main loop iteration checks intersection with
voxels on the scene (notice that there is distance limit).
Voxel is picked according to the center of the screen, not to the mouse
position.

Verification: Build, run, fly and click. Notice how fun can be changing
everything into bedrock. No crosshair, but You can put finger on the middle
of the screen - works wonders.

Change-Id: I8b9bbab14f5813ed8cd64d8be7abf9ec52b2cf5d